+++ BuildingWithParking.h
@@ -12,7 +12,7 @@ public:
 
     virtual void provideService();
 
-    virtual int getCost();
+    virtual float getCost();
 
 private:
